+++ app/views/playground.vue
@@ -19,7 +19,7 @@
 				</span>
 				<span class="dirty-badge" :class="{dirty: engraverDirty}"></span>
 				<button @click="engrave" :class="{working: engraving}" style="zoom: 160%" title="engrave">&#x1f3bc;</button>
-				<button :disabled="!sheetDocument" @click="exportScore">.json</button>
+				<button :disabled="!sheetDocument" @click="exportScore">json</button>
 			</fieldset>
 			<fieldset>
 				<BoolStoreInput v-show="false" v-model="tokenizeStaff" sessionKey="lotus-tokenizeStaff" />